# unfazed/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export interface TherapistRegistration {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export interface TherapistLogin {
  email: string;
  /** @minLength 1 */
  password: string;
}

export type TherapistServiceInputDurationMinutes = typeof TherapistServiceInputDurationMinutes[keyof typeof TherapistServiceInputDurationMinutes];


export const TherapistServiceInputDurationMinutes = {
  NUMBER_30: 30,
  NUMBER_45: 45,
  NUMBER_60: 60,
  NUMBER_90: 90,
} as const;

export interface TherapistServiceInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  title: string;
  /** @maxLength 500 */
  description: string;
  durationMinutes: TherapistServiceInputDurationMinutes;
  /** @minimum 0 */
  feeInr: number;
}

export interface TherapistProfileUpdate {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name?: string;
  /**
     * @minLength 2
     * @maxLength 48
     * @pattern ^[a-z0-9]+(?:-[a-z0-9]+)*$
     */
  slug?: string;
  /** @maxLength 3000 */
  bio?: string;
  /**
     * @maxItems 12
     * @items.minLength 2
     * @items.maxLength 80
     */
  specializations?: string[];
  /**
     * @maxItems 12
     * @items.minLength 2
     * @items.maxLength 60
     */
  languages?: string[];
  /** @maxItems 12 */
  services?: TherapistServiceInput[];
}

export type TherapistService = TherapistServiceInput & {
  id: string;
};

export interface Therapist {
  id: string;
  name: string;
  email: string;
  slug: string;
  bio: string;
  specializations: string[];
  languages: string[];
  services: TherapistService[];
}

export interface PublicTherapistProfile {
  name: string;
  slug: string;
  bio: string;
  specializations: string[];
  languages: string[];
  services: TherapistService[];
}

export interface AuthSession {
  therapist: Therapist;
}