# unfazed/lib/api-zod/src/generated/types/publicTherapistProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TherapistService } from './therapistService';

export interface PublicTherapistProfile {
  name: string;
  slug: string;
  bio: string;
  specializations: string[];
  languages: string[];
  services: TherapistService[];
}