# unfazed/lib/api-zod/src/generated/types/therapist.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TherapistService } from './therapistService';

export interface Therapist {
  id: string;
  name: string;
  email: string;
  slug: string;
  bio: string;
  specializations: string[];
  languages: string[];
  services: TherapistService[];
}