# unfazed/lib/api-zod/src/generated/types/therapistLogin.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TherapistLogin {
  email: string;
  /** @minLength 1 */
  password: string;
}