# unfazed/lib/api-zod/src/generated/types/therapistProfileUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TherapistServiceInput } from './therapistServiceInput';

export interface TherapistProfileUpdate {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name?: string;
  /**
     * @minLength 2
     * @maxLength 48
     * @pattern ^[a-z0-9]+(?:-[a-z0-9]+)*$
     */
  slug?: string;
  /** @maxLength 3000 */
  bio?: string;
  /**
     * @maxItems 12
     * @items.minLength 2
     * @items.maxLength 80
     */
  specializations?: string[];
  /**
     * @maxItems 12
     * @items.minLength 2
     * @items.maxLength 60
     */
  languages?: string[];
  /** @maxItems 12 */
  services?: TherapistServiceInput[];
}