# unfazed/lib/api-zod/src/generated/types/therapistRegistration.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TherapistRegistration {
  /**
     * @minLength 2
     * @maxLength 100
     */
  name: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}