# unfazed/lib/api-zod/src/generated/types/therapistService.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TherapistServiceInput } from './therapistServiceInput';

export type TherapistService = TherapistServiceInput & {
  id: string;
};