# unfazed/lib/api-zod/src/generated/types/therapistServiceInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TherapistServiceInputDurationMinutes } from './therapistServiceInputDurationMinutes';

export interface TherapistServiceInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  title: string;
  /** @maxLength 500 */
  description: string;
  durationMinutes: TherapistServiceInputDurationMinutes;
  /** @minimum 0 */
  feeInr: number;
}